+++ src/gcp/cloudfunctionsv2.ts
@@ -285,7 +285,7 @@ export async function getFunction(
  */
 export async function listFunctions(projectId: string, region: string): Promise<CloudFunction[]> {
   const res = await listFunctionsInternal(projectId, region);
-  if (res.unreachable!.includes(region)) {
+  if (res.unreachable.includes(region)) {
     throw new FirebaseError(`Cloud Functions region ${region} is unavailable`);
   }
   return res.functions;
@@ -333,9 +333,16 @@ async function listFunctionsInternal(
 export async function updateFunction(
   cloudFunction: Omit<CloudFunction, OutputOnlyFields>
 ): Promise<Operation> {
+  // Keys in labels and environmentVariables are user defined, so we don't recurse
+  // for field masks.
+  const fieldMasks = proto.fieldMasks(
+    cloudFunction,
+    /* doNotRecurseIn...=*/ "labels",
+    "serviceConfig.environmentVariables"
+  );
   try {
     const queryParams = {
-      updateMask: proto.fieldMasks(cloudFunction).join(","),
+      updateMask: fieldMasks.join(","),
     };
     const res = await client.patch<typeof cloudFunction, Operation>(
       cloudFunction.name,
@@ -446,6 +453,14 @@ export function functionFromEndpoint(endpoint: backend.Endpoint, source: Storage
     if (endpoint.eventTrigger.retry) {
       logger.warn("Cannot set a retry policy on Cloud Function", endpoint.id);
     }
+    // By default, Functions Framework in GCFv2 opts to downcast incoming cloudevent messages to legacy formats.
+    // Since Firebase Functions SDK expects messages in cloudevent format, we set FUNCTION_SIGNATURE_TYPE to tell
+    // Functions Framework to disable downcast before passing the cloudevent message to function handler.
+    // See https://github.com/GoogleCloudPlatform/functions-framework-nodejs/blob/master/README.md#configure-the-functions-
+    gcfFunction.serviceConfig.environmentVariables = {
+      ...gcfFunction.serviceConfig.environmentVariables,
+      FUNCTION_SIGNATURE_TYPE: "cloudevent",
+    };
   } else if (backend.isScheduleTriggered(endpoint)) {
     // trigger type defaults to HTTPS.
     gcfFunction.labels = { ...gcfFunction.labels, "deployment-scheduled": "true" };
@@ -472,7 +487,7 @@ export function endpointFromFunction(gcfFunction: CloudFunction): backend.Endpoi
   } else if (gcfFunction.eventTrigger) {
     trigger = {
       eventTrigger: {
-        eventType: gcfFunction.eventTrigger!.eventType,
+        eventType: gcfFunction.eventTrigger.eventType,
         eventFilters: {},
         retry: false,
       },