+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-- Fixes bug where updating gen 2 pubsub functions always failed (#4198).
 - Updates reserved environment variables for CF3 to include 'EVENTARC_CLOUD_EVENT_SOURCE' (#4196).
 - Fixes arg order for `firebase emulators:start --only storage` (#4195).
+- Fixes bug where environment variable for gen 2 functions weren't updated on deploy (#4209).
+- Fixes an issue in the storage emulator where a file upload would trigger functions with a metadata update handler (#4213).
 - Fixes bug where securityLevel is overwritten on https function re-deploys (#4208).